+++ flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.5+2
+
+- Add between rounds to example
+  
 ## 0.1.5+1
 
 - Add linux folder to example and also remove print statement